requirements.txt
faiss-cpu
gradio
IPython
libcst
markdown
numba
numpy
openai
pandas
pre_commit
PyGithub
python_minifier
scikit-learn
scipy
sentence-transformers
tiktoken
torch
